+++ docs/changelog.rst
@@ -21,7 +21,7 @@ Fixes
 -----
 
 * Recipe containing a MultiplotTask does not contain datasets as dicts
-
+* PlotTask with automatically generated filenames and >1 datasets writes correct filenames to figure record in recipe
 
 Version 0.6.1
 =============